Update of telescope data merger for VXD: sync check with timestamps, improved display logging.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@9396 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ eutel/merge/include/CircularTools.h
@@ -20,7 +20,7 @@ namespace CIRC {
    * but are expected to work with sets of small range compared to the tag range.
    */
 
-  typedef short int tag_type;
+  typedef int tag_type;
 
   enum {
     c_tagLowerBound = 0,
